+++ docs/datasette-publish-now-help.txt
@@ -1,31 +1,23 @@
-$ datasette publish --help
+$ datasette publish now --help
 
-Usage: datasette publish [OPTIONS] PUBLISHER [FILES]...
-
-  Publish specified SQLite database files to the internet along with a datasette API.
-
-  Options for PUBLISHER:     * 'now' - You must have Zeit Now installed:
-  https://zeit.co/now     * 'heroku' - You must have Heroku installed:
-  https://cli.heroku.com/
-
-  Example usage: datasette publish now my-database.db
+Usage: datasette publish now [OPTIONS] [FILES]...
 
 Options:
-  -n, --name TEXT           Application name to use when deploying
   -m, --metadata FILENAME   Path to JSON file containing metadata to publish
   --extra-options TEXT      Extra options to pass to datasette serve
-  --force                   Pass --force option to now
   --branch TEXT             Install datasette from a GitHub branch e.g. master
-  --token TEXT              Auth token to use for deploy (Now only)
   --template-dir DIRECTORY  Path to directory containing custom templates
   --plugins-dir DIRECTORY   Path to directory containing custom plugins
   --static STATIC MOUNT     mountpoint:path-to-directory for serving static files
   --install TEXT            Additional packages (e.g. plugins) to install
-  --spatialite              Enable SpatialLite extension
   --version-note TEXT       Additional note to show on /-/versions
   --title TEXT              Title for metadata
   --license TEXT            License label for metadata
   --license_url TEXT        License URL for metadata
   --source TEXT             Source label for metadata
   --source_url TEXT         Source URL for metadata
+  -n, --name TEXT           Application name to use when deploying
+  --force                   Pass --force option to now
+  --token TEXT              Auth token to use for deploy (Now only)
+  --spatialite              Enable SpatialLite extension
   --help                    Show this message and exit.
